2nd stage

+++ src/main.rs
@@ -19,7 +19,7 @@ fn decode_bencoded_value(encoded_value: &str) -> serde_json::Value {
         let len = encoded_value.len();
 
         match  first_char {
-          'i' => serde_json::Value::Number(encoded_value[1..len-1].parse::<i32>().unwrap().into()),
+          'i' => serde_json::Value::Number(encoded_value[1..len-1].parse::<i64>().unwrap().into()),
           _ => panic!("Unhandled encoded value: {}", encoded_value)
